+++ lib/src/ast/sass.dart
@@ -17,6 +17,7 @@ export 'sass/expression/list.dart';
 export 'sass/expression/map.dart';
 export 'sass/expression/null.dart';
 export 'sass/expression/number.dart';
+export 'sass/expression/parenthesized.dart';
 export 'sass/expression/selector.dart';
 export 'sass/expression/string.dart';
 export 'sass/expression/unary_operation.dart';